purged last instances of rid

+++ mockingbird/src/main/kotlin/com/leftindust/mockingbird/dao/RecordDao.kt
@@ -5,6 +5,6 @@ import com.leftindust.mockingbird.dao.entity.MediqRecord
 import com.leftindust.mockingbird.extensions.CustomResult
 
 interface RecordDao {
-    suspend fun getRecordByRecordId(rid: Int, requester: MediqToken): CustomResult<MediqRecord, OrmFailureReason>
+    suspend fun getRecordByRecordId(rid: Long, requester: MediqToken): CustomResult<MediqRecord, OrmFailureReason>
     suspend fun getRecordsByPatientPid(pid: Long, requester: MediqToken): CustomResult<List<MediqRecord>, OrmFailureReason>
 }
